arm/aspeed: Ensure all IP will be reset on reboot
When the watchdog counts down the processor may optionally reset all of
it's peripherals. An issue was observed where the Flash controller was
placed in 4b mode and the system would not come back up.

Set it to reset all peripherals (except the SDRAM and AHB bridge), which
is the default, to ensure that everything is reset to ensure the system
can reboot.

Signed-off-by: Joel Stanley <joel@jms.id.au>