arm: aspeed: Remove include/machine directory
Remove the need for the include/mach-aspeed directory like the
other platforms included in the multiplatform kernel.  That is,
move the include files to the machine directory with their users.

Using git grep shows exactly one use of ast2400.h in the tree.
This reduces our diff to upstream, removing the hack to allow
the include file to be found.  I decided not to merge the file
as the contents will eventually be encoded in the device tree.

Signed-off-by: Milton Miller <miltonm@us.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>