Merge tag 'aspeed-for-4.7-arch' of https://github.com/shenki/linux into next/soc
Merge "aspeed arch tag for 4.7" from Joel Stanley <joel@jms.id.au>:

This is the initial pull request for the Aspeed BMC SoCs.

I put 4.7 in the subject for this and the subsequent requests but if it's a bit
late in the cycle then I understand.

Following review I got rid of the board file by pushing the functionality out
to a watchdog and clock driver, so it's just the Kconfig bits. I've also added
myself to maintainers for the Aspeed arch and drivers as I intend on looking
after them.

* tag 'aspeed-for-4.7-arch' of https://github.com/shenki/linux:
  arm: Add Aspeed machine