aspeed_defconfig: Enable SHMEM and TMPFS
Without enabling these options we get strange behaviour under qemu,
where it appears uevents aren't generated to create device nodes for
ttyS{0,4,5}. As a consequence we don't receive a getty on the console.

Cédric narrowed down the list of options initially at [1],
and with a patch to the qemu timer model[2] we can retain the
CONFIG_NO_HZ_IDLE=y configuration.

[1] https://github.com/openbmc/qemu/issues/4
[2] https://lists.ozlabs.org/pipermail/openbmc/2016-May/003209.html

Cc: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>