net/ncsi: Enable request timer before sending command
The response is received upon the sent command before the request's
timer is enabled under extreme situation. The response packet won't
be processed as the corresponding request (request packet and timer)
isn't complete. It leads to NCSI failure - no package detected.

This fixes the issue by enabling the timer before sending the command.
The comments about the timer has been replaced by that in upstream
kernel.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>