arm/aspeed: Reset AHB on boot
The Aspeed u-boot port from the Aspeed SDK performed I2C initialisation.
Part of this appears to require a AHB reset in order to successfully
access the I2C registers.

So we can boot the stable kernels on systems with new u-boot we add this
to the stable kernel tree.

Signed-off-by: Joel Stanley <joel@jms.id.au>