aspeed: fix the debugfs write operation
It was incompatible with the seq_file used in the read so the read
operation was changed also. Now you can write to the debug file with:

  # echo 0x1E630004=0x3DBB2081 > /sys/kernel/debug/aspeed/1e630000.spi
  # cat /sys/kernel/debug/aspeed/1e630000.spi
  0x1E630000: 0x00000002
  0x1E630004: 0x3DBB2081
  0x1E630008: 0x00000000
  0x1E63000C: 0x00000000
  0x1E630010: 0x00000000
  0x1E630014: 0x00000000

Some basic checks are done to not ruin the kernel but anyhow, this is
not for upstream. It is just a useful devmem replacement.

Signed-off-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Joel Stanley <joel@jms.id.au>