arm: aspeed: Remove setting of SCU90 from common
This hack enabled I2C buses 3 to 8 inclusive:

   Enable I2C8 function pins: 0x1
   Enable I2C7 function pins: 0x1
   Enable I2C6 function pins: 0x1
   Enable I2C5 function pins: 0x1
   Enable I2C4 function pins: 0x1
   Enable I2C3 function pins: 0x1

Bit 3 set the USB1.1 Host port 2 function, which is not yet supported in our
kernel.

Bits 13 and 15 disabled the RGMII/RMII 1 and 2 RX pin internal pull down which
is the default state.

Therefore all functionality is present in the device tree and this can be
removed.

Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>