arm: aspeed: Remove UART muxing hacks
These were added during initial bringup. None of these pins are used by
OpenBMC for the functions they're being muxed to (UART4 nor the VGA
output) so there are no device tree additions required to cover for
the removal.

We do use UART5, which is an exclusive function of the pins it sits on,
so again no pinmux configuration is required.

Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>