arm: aspeed: zaius: Disable LPC reset for UART1
Currently, UART1 on Zaius BMC is unusable until brought out of reset by
powering the host on. In that default reset state, ttyS0 can still be
opened and UART1 silently drops bytes, which is not obviously expected
behavior.

Clear the LPC block control bit that enables LPCRST# as a reset source
for UART1, making UART1 immediately useful in that it won't discard
bytes.

Signed-off-by: Xo Wang <xow@google.com>
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
[fixed comment formatting]
Signed-off-by: Joel Stanley <joel@jms.id.au>