drivers/fsi: Clock slave prior to master command
If the FSI bus has been idle for some period it is necessary to
wake up the slave so its ready for a new master initiated
commands. This is done by clocking out a series of 0's on the
SDA line immediately before any command.

Signed-off-by: Christopher Bostic <cbostic@linux.vnet.ibm.com>
[applied as a temporary measure until correct bus recovery and error
handling is added to the core]
Signed-off-by: Joel Stanley <joel@jms.id.au>