arm: aspeed: romulus: Set PNOR SPI address mapping
The PNOR SPI address mapping is the same as Witherspoon.

Even though Romulus only has the one flash chip we want to map (on CE0),
we need to move the CE1 mapping/window out of the way or else the
controller is unhappy.

The upstream aspeed smc driver will gain support for setting these
registers in the future so we can drop these hacks.

Signed-off-by: Lei YU <mine260309@gmail.com>
Acked-by: Andrew Jeffery <andrew@aj.id.au>
[Joel: add some content to the commit message]
Signed-off-by: Joel Stanley <joel@jms.id.au>