ARM: dts: aspeed: zaius: Enable NC-SI mux
The Zaius motherboard has muxes for the RMII lines from the BMC to:
  - BCM5719 LAN on motherboard (LOM)
  - OCP mezzanine connector
  - HDMI form factor connector

Assert the active-low enable line for the muxes. Due to the AST2500
default internal pull-down resistors, the muxes selection lines will
be low. This implicitly selects the LOM for NC-SI.

The NC-SI interface will show up as network interface eth0.

Signed-off-by: Xo Wang <xow@google.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>