ARM: dts: aspeed: Add Palmetto hogs
This is the ugly hog patch Andrew cooked up several months ago when we
had issues with the post-1.0 OpenBMC branch on Palmetto.

The idea was these hogs were to make it into userspace. No one has taken
on that work, so we will do it in the kernel for now.

There are also some pinmux hogs, where the default mode of the IP block
is configured.

OpenBMC-Staging-Count: 2
Signed-off-by: Joel Stanley <joel@jms.id.au>