ARM: dts: Disable UCD90160 in the Witherspoon dts
This is a work-around for the lack of PGOOD issue:

  https://github.com/openbmc/openbmc/issues/832

Binding the UCD9000 driver to the device (and possibly the sequence of
unbinding, poking the UCD90160 via i2cset, then re-binding the UCD9000
driver) appears to have some negative effects.  Instrumenting the
transactions with reads of the PMBus status tree lead to the following
observations under "normal" operation.

Reduce our problems by communicating with the UCD90160 as little as
possible.  With the associated userspace change, the only communication
will be applying the VCS work-around.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Tested-by: Uma Yadlapati <yadlapat@us.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>