tty: aspeed-vuart: Add IRQF_SHARED to flags
All of the LPC devices share an IRQ line. When other LPC devices (such
as the BT) have enabled IRQs we receive them unexpectedly in the vuart
driver before it is ready.

And the IRQF_SHARED flag to the vuart driver so we ignore interrupts
until we are ready to handle them.

Reported-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Joel Stanley <joel@jms.id.au>