mtd: spi-nor: aspeed: configure chip window on AHB bus
The segment registers of the SMC controller provide a way to configure
the mapping windows of the chips on the AHB bus. The settings are
required to be correct when the controller operates in Command mode,
which is the case for DMAs and the LPC mapping.

This tries to set the segment registers of each chip depending on the
size of the flash device and depending on the previous segment
settings, in order to have a contiguous window across multiple chip.

Unfortunately, the AST2500 SPI controller has a bug and it is not
possible to configure a full 128MB window for a chip of the same
size. The window size needs to be restricted to 120MB. This issue only
applies to CE0.

Sample kernel dmesg on a Witherspoon :

  aspeed-smc 1e620000.flash-controller: Using IRQ 17
  aspeed-smc 1e620000.flash-controller: DMA support disabled.
  aspeed-smc 1e620000.flash-controller: mx25l25635e (32768 Kbytes)
  aspeed-smc 1e620000.flash-controller: CE0 window [ 0x20000000- 0x22000000 ] 32MB
  aspeed-smc 1e620000.flash-controller: CE1 window [ 0x22000000- 0x2a000000 ] 128MB
  5 ofpart partitions found on MTD device bmc
  Creating 5 MTD partitions on "bmc":
  0x000000000000-0x000000060000 : "u-boot"
  0x000000060000-0x000000080000 : "u-boot-env"
  0x000000080000-0x0000004c0000 : "kernel"
  0x0000004c0000-0x000001c00000 : "rofs"
  0x000001c00000-0x000002000000 : "rwfs"
  aspeed-smc 1e620000.flash-controller: mx25l25635e (32768 Kbytes)
  aspeed-smc 1e620000.flash-controller: CE1 window [ 0x22000000- 0x24000000 ] 32MB
  aspeed-smc 1e620000.flash-controller: CE2 window [ 0x24000000- 0x2c000000 ] 128MB
  aspeed-smc 1e630000.flash-controller: mx66l1g45g (131072 Kbytes)
  aspeed-smc 1e630000.flash-controller: CE0 window resized to 120MB (AST2500 HW quirk)
  aspeed-smc 1e630000.flash-controller: CE0 window [ 0x30000000- 0x37800000 ] 120MB
  aspeed-smc 1e630000.flash-controller: CE0 window too small for chip 128MB
  aspeed-smc 1e630000.flash-controller: CE1 window [ 0x37800000- 0x38000000 ] 8MB

OpenBMC-Staging-Count: 1
Signed-off-by: Cédric Le Goater <clg@kaod.org>
Acked-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>